% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categories.R
\docType{data}
\name{categories}
\alias{categories}
\title{COG Information data
In the functional annotation dataset (og_categories), a COG's functional category is represented by a single letter coding. This dataset
translates the single letter encoding into a more detailed description}
\format{
A dataframe with 1048575 rows and 3 variables
\describe{
\item{group1}{reference ID for the first COG of the pair}
\item{group2}{reference ID for the second COG of the pair}
\item{associationscore}{association score between the two COGs}
}
}
\source{
\url{https://stringdb-static.org/cgi/download.pl?sessionId=A8sx4Z24Onr2}
}
\usage{
data(categories)
}
\description{
COG Information data
In the functional annotation dataset (og_categories), a COG's functional category is represented by a single letter coding. This dataset
translates the single letter encoding into a more detailed description
}
\keyword{datasets}
