% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coglinker.R
\name{cog_linker}
\alias{cog_linker}
\title{The Annotation Function}
\usage{
cog_linker(groups, benchmark)
}
\arguments{
\item{groups}{A character vector that contains a list of OGs, benchmark for association score}
}
\value{
dataframe containing each pair of OGs with its bitscore
}
\description{
This function, when fed a character vector of ortholog groups, will return the bit scores of the COG group associations
}
\examples{
\dontrun {
get_orthology([Group X])
}
The Group X pairs with Group Y with a bit score of 248
}
\keyword{annotation}
\keyword{functional}
