% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cog_links.R
\docType{data}
\name{cog_links}
\alias{cog_links}
\title{COG Information data
This dataset shows the association between two COGs, demonstrating the likelihood of functional similarity}
\format{
A dataframe with 1048575 rows and 3 variables
\describe{
\item{group1}{reference ID for the first COG of the pair}
\item{group2}{reference ID for the second COG of the pair}
\item{associationscore}{association score between the two COGs}
}
}
\source{
\url{https://stringdb-static.org/cgi/download.pl?sessionId=A8sx4Z24Onr2}
}
\usage{
data(cog_links)
}
\description{
COG Information data
This dataset shows the association between two COGs, demonstrating the likelihood of functional similarity
}
\keyword{datasets}
