% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CogLinker.R
\name{CogLinker}
\alias{CogLinker}
\title{The Annotation Function}
\usage{
CogLinker(groups, benchmark)
}
\arguments{
\item{groups}{A character vector that contains a list of OGs, benchmark for association score}
}
\value{
dataframe containing each pair of OGs with its bitscore
}
\description{
This function, when fed a character vector of ortholog groups, will return the bit scores of the COG group associations
}
\examples{
\dontrun {
CogLinker([Group X])
}
The Group X does not associate with any COG
}
\keyword{annotation}
\keyword{functional}
