% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{functional_annotation}
\alias{functional_annotation}
\title{The Annotation Function}
\usage{
functional_annotation(classifiers)
}
\arguments{
\item{classifiers}{A list that contains a sequence of orthologs}
}
\value{
dataframe containing the functional annotations for each ortholog group
}
\description{
This function retrieves the corresponding ortholog groups for the input set of proteins, by retrieving
their ortholog groups and that orthologs members (derived from EGGNOG Database)
}
\examples{
\dontrun {
get_orthology([Group X])
}
The Group X has the functional assignment A which is associated with the functional description: DNA Repair
}
\keyword{annotation}
\keyword{functional}
