% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionalAnnotation.R
\name{FunctionalAnnotation}
\alias{FunctionalAnnotation}
\title{The Annotation Function}
\usage{
FunctionalAnnotation(classifiers)
}
\arguments{
\item{classifiers}{A list that contains a sequence of orthologs}
}
\value{
dataframe containing the functional annotations for each ortholog group
}
\description{
This function retrieves the corresponding ortholog groups for the input set of proteins, by retrieving
their ortholog groups and that orthologs members (derived from EGGNOG Database)
}
\examples{
{
FunctionalAnnotation(c("COG0001"))
}
"The group COG0001 has the functional assignment
 H which is associated with the functional description: 
 Coenzyme transport and metabolism"
}
\keyword{annotation}
\keyword{functional}
