% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthology.R
\name{get_orthology}
\alias{get_orthology}
\title{The Get Orthology Function}
\usage{
get_orthology(classifiers)
}
\arguments{
\item{classifiers}{A list that contains a sequence of protein/genes classifiers}
}
\value{
dataframe containing the orthologous groups present in the input sequence, and their respective ortholog groups
}
\description{
This function retrieves the corresponding ortholog groups for the input set of proteins, by retrieving
their ortholog groups and that orthologs members (derived from EGGNOG Database)
}
\examples{
\dontrun {
get_orthology([Gene1,Gene2,Gene3])
}
{Gene1: Group X, Gene2: Group Y, Gene3: Group Z}
}
\keyword{orthology}
