% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetOrthology.R
\name{GetOrthology}
\alias{GetOrthology}
\title{The Get Orthology Function}
\usage{
GetOrthology(classifiers)
}
\arguments{
\item{classifiers}{A list that contains a sequence of protein/genes classifiers}
}
\value{
dataframe containing the orthologous groups present in the input
sequence, and their respective ortholog groups
}
\description{
This function retrieves the corresponding ortholog groups for the input set of proteins, by retrieving
their ortholog groups and that orthologs members (derived from EGGNOG Database)
}
\examples{
{
GetOrthology(c("MK0280","MA_0581","PTH_0979"))
}
"Protein with GenBankID MK0280 belongs to COG0001"
"Protein with GenBankID MA_0581 belongs to COG0001"
"Protein with GenBankID PTH_0979 belongs to COG0001"
}
\keyword{orthology}
