% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_data}
\alias{load_data}
\title{The Load Data Function}
\usage{
load_data(directory)
}
\arguments{
\item{directory}{that the data is located in}
}
\description{
This function is used to set up the pertaining databases in the package
}
\examples{
\dontrun {
load_data()
}
The data has been loaded!


}
\keyword{load,data}
