% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OG_network.R
\name{ortholog_network}
\alias{ortholog_network}
\title{The OG Network Function}
\usage{
ortholog_network(proteins, benchmark)
}
\arguments{
\item{classifiers}{A list that contains a sequence of NCBI/Genbank protein ids}
}
\value{
graph of interaction network
}
\description{
from beginning to end, using the query functions that were done before, this function constructs a protein interaction network
that is orthog specific. Each node is a COG group, with information on the functional annotation  and the input proteins present in each ortholog group. there
are lines connecting interaction scores
}
\examples{
\dontrun {
protein_network([Protein X])
}

}
\keyword{interaction}
\keyword{network}
