% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrthologData.R
\docType{data}
\name{OrthologsData}
\alias{OrthologsData}
\title{COG Information data}
\format{
A dataframe with 3455853 rows and 11 variables
\describe{
\item{GenBankID}{reference ID for protein in GenBank database}
\item{NCBI_ID}{reference ID of protein in NCBI database}
\item{ProteinID}{association score between the two COGs}
\item{Proteinlenght}{length of protein}
\item{COGID}{reference ID of the COG the protein belongs to}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/research/cog-project/}
}
\usage{
data(OrthologsInfo)
}
\description{
Given the various reference IDs for a certain protein, this dataset maps the protein to the correct Cluster of Orthologous Group (COG),
a classification that represents function categorization
}
\keyword{datasets}
