% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_network.R
\name{protein_network}
\alias{protein_network}
\title{The Protein Network Function}
\usage{
protein_network(proteins, benchmark)
}
\arguments{
\item{classifiers}{A list that contains a sequence of NCBI/Genbank protein ids}
}
\value{
graph of interaction network
}
\description{
from beginning to end, using the query functions that were done before, this function constructs a protein interaction network
that is protein-specific-each node is a protein, with information on the functional annotation and the ortholog group. there
are lines connecting interactoin scores
}
\examples{
\dontrun {
protein_network([Protein X])
}

}
\keyword{interaction}
\keyword{network}
