% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMicroMetaGraph.R
\name{runMicroMetaGraph}
\alias{runMicroMetaGraph}
\title{Launch Shiny App For Package MicroMetaGraph}
\usage{
runMicroMetaGraph()
}
\value{
No return value but open up a shiny page.
}
\description{
A function that launches the shiny app for this package.
The shiny app permit to perform clustering using mixtures
of MPLN via variational-EM. Model selection is performed
using AIC, AIC3, BIC and ICL.
}
\examples{
\dontrun{
runMicroMetaGraph()
}

}
\author{
Emily So , \email{emily.so@mail.utoronto.ca}
}
