% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CogLinker.R
\name{CogLinker}
\alias{CogLinker}
\title{The Annotation Function}
\usage{
CogLinker(groups, benchmark)
}
\arguments{
\item{groups}{A character vector that contains a list of OGs, benchmark for association score}
}
\value{
dataframe containing each pair of OGs with its bitscore
}
\description{
This function, when fed a character vector of ortholog groups, will return the bit scores of the other COG groups in the datasets that it interacts with
}
\examples{
\dontrun {
CogLinker(c("COG0001","COG0002"),100)
}
"Classifier COG0002 connects with the group COG0001"
}
\keyword{annotation}
\keyword{functional}
