% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrthologNetwork.R
\name{OrthologNetwork}
\alias{OrthologNetwork}
\title{The OG Network Function}
\usage{
OrthologNetwork(proteins, benchmark = 0)
}
\arguments{
\item{proteins}{A list that contains a sequence of NCBI/Genbank protein ids}

\item{benchmark}{Set benchmarks for association scores}
}
\value{
graph of interaction network
}
\description{
from beginning to end, using the query functions that were done before, this function constructs a protein interaction network
that is orthog specific. Each node is a COG group, with information on the functional annotation  and the input proteins present in each ortholog group. there
are lines connecting interaction scores
}
\keyword{interaction}
\keyword{network}
