% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProteinNetwork.R
\name{ProteinNetwork}
\alias{ProteinNetwork}
\title{The Protein Network Function}
\usage{
ProteinNetwork(proteins, benchmark = 0)
}
\arguments{
\item{proteins}{A list that contains a sequence of NCBI/Genbank protein ids}

\item{benchmark}{Minimum benchmark for associaton score}
}
\value{
graph of interaction network
}
\description{
from beginning to end, using the query functions that were done before, this function constructs a protein interaction network
that is protein-specific-each node is a protein, with information on the functional annotation and the ortholog group. there
are lines connecting interactoin scores
}
\keyword{interaction}
\keyword{network}
